+++ tests/api/BoreholeLockServiceTest.cs
@@ -106,6 +106,50 @@ public class BoreholeLockServiceTest
         Assert.AreEqual(false, await boreholeLockService.IsBoreholeLockedAsync(borehole.Id, AdminSubjectId));
     }
 
+    [TestMethod]
+    public async Task IsUserLackingPermissionsWithSufficientPermissions()
+    {
+        var users = await context.Users.Include(s => s.WorkgroupRoles).ToListAsync().ConfigureAwait(false);
+        var user = users.FirstOrDefault(u => u.FirstName.Equals("editor", StringComparison.OrdinalIgnoreCase));
+
+        var boreholes = await context.Boreholes.Include(b => b.Workflows).ToListAsync();
+        boreholes = boreholes.Where(x => x.Workflows.Count > 0 && x.Workflows.Any(w => w.Role == Role.Editor)).Take(2).ToList();
+
+        var result = boreholeLockService.IsUserLackingPermissions(boreholes, user);
+        Assert.IsFalse(result);
+    }
+
+    [TestMethod]
+    public async Task IsUserLackingPermissionsWithNotMatchingWorkgroupId()
+    {
+        var users = await context.Users.Include(s => s.WorkgroupRoles).ToListAsync().ConfigureAwait(false);
+        var user = users.FirstOrDefault(u => u.FirstName.Equals("editor", StringComparison.OrdinalIgnoreCase));
+
+        var boreholes = await context.Boreholes.Include(b => b.Workflows).ToListAsync();
+        boreholes = boreholes.Where(x => x.Workflows.Count > 0 && x.Workflows.Any(w => w.Role == Role.Editor)).Take(2).ToList();
+
+        // Update workgroupId of first borehole to simulate a workgroup mismatch
+        boreholes[0].WorkgroupId = 2;
+
+        var result = boreholeLockService.IsUserLackingPermissions(boreholes, user);
+        Assert.IsTrue(result);
+    }
+
+    [TestMethod]
+    public async Task IsUserLackingPermissionsWithNotMatchingWorkflowRole()
+    {
+        // Retrieve a user with the 'Viewer' role, which lacks sufficient permissions to manage boreholes
+        // that are associated with workflows requiring the 'Editor'(Role.Editor) role.
+        var users = await context.Users.Include(s => s.WorkgroupRoles).ToListAsync().ConfigureAwait(false);
+        var user = users.FirstOrDefault(u => u.FirstName.Equals("viewer", StringComparison.OrdinalIgnoreCase));
+
+        var boreholes = await context.Boreholes.Include(b => b.Workflows).ToListAsync();
+        boreholes = boreholes.Where(x => x.Workflows.Count > 0 && x.Workflows.Any(w => w.Role == Role.Editor)).Take(2).ToList();
+
+        var result = boreholeLockService.IsUserLackingPermissions(boreholes, user);
+        Assert.IsTrue(result);
+    }
+
     private Borehole GetLockedBorehole(bool lockedByAdmin)
     {
         bool LockedCondition(Borehole borehole) => lockedByAdmin ? borehole.LockedById == AdminUserId : borehole.LockedById != AdminUserId;