+++ CHANGELOG.md
@@ -15,6 +15,7 @@
 ### Fixed
 
 - When hovering over the main side navigation icons, an incorrect background color was displayed for the hovered icon.
+- Before export check if the user has the correct role and is in the same workgroup as the borehole.
 
 ## v2.1.1052 - 2025-02-04